Changed the link manager and telemetry scheduler so that the rate filter values that are unused are grayed out and unselectable. Corrected the table type editor input type column so that grayed out types are unselectable. This includes the addition of two new PostgreSQL functions
Updated JTableHandler to replace the TABLE_CHANGE_EVENT property change
event with a placeholder method. The method is called with invokeLater()
to ensure all other table changes are handled prior to calling the
placeholder method. Changed the various editors to override the
placeholder instead of listening for the TABLE_CHANGE_EVENT