Added auto-complete to the database host text fill.  Added tool tips showing the complete script path in the script manager and script execution associations lists.  Changed the script manager and script execution associations lists to show only the script name even for associations created in other OS's.  .jar file NOT updated (in work) Adding update to associations when a table is renamed or deleted.
Updated table and table type editors so that change indicators
(asterisks) are moved from all tabs when the 'Store all' command is
performed.  Corrected update of table editor tab names if a prototype
table is renamed.  Updated PostgreSQL command used to alter data table
names in the internal tables. 
Added updating of script associations in the __associations table when a
table is renamed or deleted.  Corrected handling parsing an association
when the association has no tables.

Corrected isTargetInTablePath() so that it stops searching only when a
match is found. Corrected the script manager to highlight all tables in
the tree that are in a selected association. Incorporated a check for
bit-packed variables when updating internal tables following a change in
a variable's bit length. 

Changed table tree to display table arrays in correct order (by array
dimension indices). Updated PostgreSQL function to account for a macro
in the array size where the macro value is blank. Corrected setting of
keyboard mnemonics for menu components. 
(in work) Corrected inital height for dialogs containing tables

Corrected random height sizing of the data type and macro editors.
Corrected exception due to invalid column index in data type and macro
editors. Changed data type and macro editors to allow deletion of an
empty row. Added the table tree and script associations in the script
manager to a split pane. Changed the dialog handler's check box and
radio button array methods to set the number of initial rows.

Added web access command to authenticate user login.  Updated table
addition, modification, and deletion methods with respect to updating
references to structures in the internal tables to cover edge cases,
changes of a table from being a root to being a child table, updates to
the links and telemetry scheduler tables, etc.  Updated user's guide and
.jar file.