Updated table import: 1) Disable table editor dialog menu commands until import complete. 2) Limit the number of tables added to an editor dialog (number is user-selectable). Create a new editor when the limit is reached. 3) Show newly created table types in the table type manager (if open). 4) Don't highlight the imported cells if the table is created for importing (retain cell highlighting if importing into an existing table). 5) Corrected XTCE import so that all prototype tables are created.
Rewrote setControlsEnabled() methods in the table type and table editors
so that the menu bar items are grayed out when the editor is disabled.
Updated so that the recent projects and recent tables menu items are
disabled when the other menu items are disabled. Changed the number of
tables added to an editor dialog during import to be based on the number
of tab rows. Updated the database functions to use a common temporary
table that is persistent for a session (versus creating/dropping the
temporary table per function call). Added a halt dialog with progress to
the import operation (in work). Updated CCDD.jar.