*** EXTENSIVE ALTERATIONS HAVE BEEN INTRODUCED THAT SUPPORT DISPLAY OF CHECK BOXES IN DATA TABLE CELLS FOR A BOOLEAN INPUT TYPE, TO ALLOW CREATION OF NEW INPUT TYPES, AND TO CREATE/DISPLAY A NEW LIST SELECTION INPUT TYPE. ONLY BASIC TESTING OF THESE CHANGES HAS BEEN PERFORMED TO DATE ***
Updated user's guide in regards to the padding dialog. Changed the EDS
and XTCE import so that the EDS/XTCE table types are not created if
these already exist. Corrected import so that the Halt dialog is
automatically removed when the export completes. Changed the data tables
so that columns having a boolean input type are displayed as check
boxes. Added information to the user's guide regarding the use of
try-catch type statements in scripts in order to enhance error
information in the log entry. Changed message ID assignment so that it
ignores data fields that aren't applicable. Added data field change
check for an applicability update. Added the table path to the text in
the table editor cell content error message dialogs. Added capability to
handle custom input types, including user-defined selection lists. Added
an input type editor which includes the capability of defining input
types that constrain the input to a set of itenms displayed in a table
cell as a combo box. Added detail as to the affected table, etc. to
error messages when deleting a macro, data type, or input type.
Corrected search_tables() function when flagging reserved characters in
the search text. Updated CCDD.jar.