*** NOTE: This version alters a project's database schema. Earlier CCDD versions will be incompatible with the updated database and vice versa ***
Changed handling of table type data fields so that inheritance by tables
is strictly enforced. Inherited fields may only be edited in the table
type editor's field editor. Adding, modifying, and deleting fields for a
type affects all tables of that type. Separator and break fields are not
inheritable (Github issue #44). Corrected table import so that the list
of all tables has newly created prototype table names added to it; this
prevents SQL commands to create a table that already exists (which would
result in an error) (Github issue #48). Corrected
buildModificationCommand() so that a change to the custom values table
value column is delimited properly to prevent issues with special
characters. Corrected variable tree creation; in table tree handler's
buildNodes() method moved the addition of the child node to the parent
node so that it occurs prior to getting the child node's full variable
path. Added a failure message to the login dialog if an invalid user
name or password is entered. Changed the exit value to always be '1' if
the application exits abnormally when executed from the command line.
Changed the table type editor Fields menu to allow four overwrite values
types (overwrite all, overwrite if value is the same, overwrite if value
is empty, and don't overwrite). Changed the patch handler to have two
sections - patches to apply prior to initializing class handlers and
patches to apply after initializing class handlers. Corrected missing
table type data fields in CSV exports. Moved the Overwrite values
sub-menu in the table type editor to the Fields menu. Updated Javadoc,
user's guide, and CCDD.jar.