Github issue #50: In CcddMain(), when the GUI is hidden, changed so that the result of the openDatabase() call and a comparison of the targeted and opened project names (in case the user specifies a project and it downmodes to the default, postgres) determines the exit status.
Github issue #52: The non-root prototypes are left out when using the
command line export.  When exporting from the dialog all tables in the
'All tables' group node in the table tree are included.  The command
line export doesn't have a tree with selected nodes; it must parse a
string of tables, so it uses a method in the script handler for this
purpose.  This method started with root (not prototype) tables, and then
added child tables.  The method was altered to include the prototypes as
well.  In addition, the script handler's method didn't automatically
include ancestors as does the export dialog's method, so this was added
in the script handler method (when using the method for exporting, not
when handling a script association).

Github issue #54: Moved the code that does the macro updates to
createTablesFromDefinitions() from importFile() since the code to store
to macros to the database was occurring prior to making the updates.

Updated Javadoc and CCDD.jat (version 1.4.57).