Ensure a Command Argument can be added to a Command
Inserting a new row in a command and selecting a command argument will
cause an error when storing the changes. This occurs in the function
modifyInternalFieldsTable when checking the data type column of the
table. However, the Command table does not have a data type column, and
an exception gets thrown.

Address this specific issue by checking if the data type column exists
before trying get the row data.

Fixes #83